+++ logprep/processor/grokker/processor.py
@@ -37,7 +37,7 @@ from zipfile import ZipFile
 from attrs import define, field, validators
 
 from logprep.abc.processor import Processor
-from logprep.processor.base.exceptions import FieldExistsWarning, ProcessingWarning
+from logprep.processor.base.exceptions import FieldExistsWarning, ProcessingWarning, ProcessingError
 from logprep.processor.grokker.rule import GrokkerRule
 from logprep.util.getter import GetterFactory
 from logprep.util.helper import add_field_to, get_dotted_field_value
@@ -69,7 +69,14 @@ class Grokker(Processor):
                 error = BaseException(f"{self.name}: missing source_field: '{dotted_field}'")
                 self._handle_warning_error(event=event, rule=rule, error=error)
                 continue
-            result = grok.match(field_value)
+            try:
+                result = grok.match(field_value)
+            except TimeoutError as error:
+                raise ProcessingError(
+                    self,
+                    f"Grok pattern timeout for source field: '{dotted_field}' in rule '{rule}', "
+                    f"the grok pattern might be too complex.",
+                ) from error
             if result is None or result == {}:
                 continue
             matches.append(True)